refactor: Refactor check_pdf_fields to return BytesIO and filename
This commit refactors the `check_pdf_fields` function to return a `io.BytesIO` object containing the filled PDF and the filename, instead of a list of `FileDataSchema` objects.

This change is necessary to facilitate the direct download of the filled PDF in the web UI. The previous implementation required further processing to convert the data into a downloadable format.