feat: Add main bucket name to constants
This commit adds the `MAIN_BUCKET_NAME` constant to the `constants.py` file. This constant stores the name of the main Google Cloud Storage bucket used by the application. This is necessary to access the bucket from different parts of the application without hardcoding the name. The value is retrieved from the environment
variable `MAIN_BUCKET_NAME`.