refactor: Remove `path` attribute from FileDataSchema in tests
The `path` attribute in the `FileDataSchema` is being removed from the test results. This is part of a broader refactoring effort to decouple the storage interface tests from specific path implementations. Removing the path assertion simplifies the tests and makes them more focused on the core functionality of the storage interface. This change does not introduce any breaking changes as it only affects test assertions.