feat: Implement delete_managed_folder in Google Storage
Implements the `delete_managed_folder` method in the Google Storage class.
This allows for the deletion of managed folders within the Google Cloud Storage bucket.
The method constructs a `DeleteManagedFolderRequest` and uses the `StorageControlClient` to delete the folder. It returns a `FolderDeleteSchema` object containing the name of the deleted folder.
Also added test lines to the bottom of the file to test delete_managed_folder functionality.