refactor: Update list_folders return type to FolderDataSchema
The `list_folders` method in the `CloudStorageInterface` was updated to return a list of `FolderDataSchema` instead of `FolderBaseSchema`. This change provides more comprehensive folder information to the caller, including potentially additional data that `FolderDataSchema` may contain compared to `FolderBaseSchema`. This change is made to prepare for upcoming UI enhancements that require more folder data.