feat: Add method to retrieve folder metadata
This commit introduces a new method, `get_folder`, to the CloudStorageService class. This method allows retrieval of folder metadata given a folder path. It leverages the underlying cloud storage's `get_folder` method to fetch and return the FolderDataSchema.
This functionality is needed to provide detailed information about specific folders in the cloud storage.