fix: Correctly access file type in _filter_files method
The code was using dictionary-style access (item["type"]) to retrieve the 'type' attribute of the `FileSchemaForFolder` object. This is incorrect because `FileSchemaForFolder` has a `type` attribute that should be accessed directly (item.type).

This change fixes a potential error and ensures the code correctly filters files based on their type.