feat: Add SubCategoryDetailSchema and is_detail param to get_subcategory
This commit introduces the `SubCategoryDetailSchema` which extends `SubCategorySchema` to include a list of `ItemSchema` objects, providing more detailed information about a subcategory.

The `get_subcategory` method in the `knowledge_category_interface` has been updated to accept an optional `is_detail` boolean parameter. When `is_detail` is True, the method returns a `SubCategoryDetailSchema` instance; otherwise, it returns a `SubCategorySchema` instance. This allows consumers to selectively request detailed or basic subcategory information.