feat: Introduce SearchQueryParameter and update find method
This commit introduces a `SearchQueryParameter` NamedTuple to represent search query parameters in a structured way. It also updates the `find` method in the `DocumentDatabase` interface to accept a list of `SearchQueryParameter` objects instead of a dictionary, providing more flexibility and type safety for search queries. Also, the limit parameter in `list` method's type is changed from `int` to `Optional[int]`.