feat: Add knowledge base entities module
This commit introduces the `src/domain/knowledge_base/entities/__init__.py` file to initialize the knowledge base entities module. This is part of the larger effort to refactor the project to follow Domain-Driven Design (DDD) principles.  This module will contain the definition of entities related to the knowledge base domain.